+++ R/expand-harmonics.R
@@ -19,12 +19,16 @@
 #' @param digits
 #' Number of digits to which each partial's MIDI pitch should be rounded.
 #'
+#' @param label_harmonics
+#' If TRUE, then the harmonics in the resulting spectrum are labelled with their harmonic numbers.
+#'
 #' @rdname expand_harmonics
 #' @export
 expand_harmonics <- function(x,
                              num_harmonics = 11L,
                              roll_off = 1,
-                             digits = 6) {
+                             digits = 6,
+                             label_harmonics = FALSE) {
   UseMethod("expand_harmonics")
 }
 
@@ -33,11 +37,13 @@ expand_harmonics <- function(x,
 expand_harmonics.sparse_fr_spectrum <- function(x,
                                                 num_harmonics = 11L,
                                                 roll_off = 1,
-                                                digits = 6) {
+                                                digits = 6,
+                                                label_harmonics = FALSE) {
   expand_harmonics(sparse_pi_spectrum(x),
                    num_harmonics = num_harmonics,
                    roll_off = roll_off,
-                   digits = digits) %>%
+                   digits = digits,
+                   label_harmonics = label_harmonics) %>%
     sparse_fr_spectrum()
 }
 
@@ -46,18 +52,21 @@ expand_harmonics.sparse_fr_spectrum <- function(x,
 expand_harmonics.sparse_pi_spectrum <- function(x,
                                                 num_harmonics = 11L,
                                                 roll_off = 1,
-                                                digits = 6) {
+                                                digits = 6,
+                                                label_harmonics = FALSE) {
   template <- pi_harmonic_template(num_harmonics, roll_off)
 
   purrr::map2(pitch(x), amp(x),
               function(pitch, amp) {
-                data.frame(
+                df <- data.frame(
                   x = pitch + template$interval,
                   y = amp * template$amplitude
                 )
+                if (label_harmonics) df$labels <- seq_along(template$interval)
+                df
               }) %>%
     collapse_summing_amplitudes(digits = digits) %>%
-    {.sparse_pi_spectrum(pitch = .$x, amplitude = .$y)}
+    {.sparse_pi_spectrum(pitch = .$x, amplitude = .$y, labels = .$labels)}
 }
 
 #' @rdname expand_harmonics
@@ -65,11 +74,13 @@ expand_harmonics.sparse_pi_spectrum <- function(x,
 expand_harmonics.pi_chord <- function(x,
                                       num_harmonics = 11L,
                                       roll_off = 1,
-                                      digits = 6) {
+                                      digits = 6,
+                                      label_harmonics = FALSE) {
   sparse_pi_spectrum(x,
                      num_harmonics = num_harmonics,
                      roll_off = roll_off,
-                     digits = digits)
+                     digits = digits,
+                     label_harmonics = label_harmonics)
 }
 
 pi_harmonic_template <- function(num_harmonics, roll_off, digits = 6) {